+++ merged-packages/bitcoin-wallet-snap/src/entities/send-flow.ts
@@ -17,6 +17,7 @@ export type ConfirmSendFormContext = {
   locale: string;
   psbt: string;
   origin?: string;
+  isMine: boolean;
 };
 
 export type SendFormContext = {